recipe: Release 28.3.0.
Note: 28.2.0 and 28.3.0 are identical. So we skipped the former. Plus
there is no changelog entry for the former so we used the latter. There
may have been issues uploading to PyPI or similar.

Additional note: 28.1.0 amounted to an adjustment of the certifi link.
As such, it seems totally unnecessary to build a conda package for it as
one could use 28.0.0 just the same.