Add a 59.x LTS branch
This is needed because multiple projects need to pin to
`setuptools < 60.0`, which was the release that enabled the
vendored copy of `distutils` inside `setuptools`. That
`distutils` copy contains backwards incompatible changes which
NumPy and SciPy don't support; other projects like Pandas and
Dask use that pin too because of the `LooseVersion` deprecation
in 60.0

See https://github.com/conda-forge/conda-forge-repodata-patches-feedstock/pull/204
for more details.

[ci skip]